ci(coverage): split into unit and CLI coverage; set thresholds (unit=85, cli=60); restructure tests into tests/unit and tests/cli; add CLI runner to tests/utils; fix paths
