+++ .github/workflows/demo.yml
@@ -23,9 +23,12 @@ jobs:
           python -m pip install -U pip
           python -m pip install -r requirements.txt -r requirements-dev.txt
           python -m pip install -e .
-      - name: Run tests with coverage (fail-under=70)
+      - name: Unit coverage (fail-under=85)
         run: |
-          pytest --cov=teds_core --cov=teds --cov-branch --cov-report=term-missing --cov-report=xml --cov-fail-under=70 -q
+          pytest tests/unit --cov=teds_core --cov=teds --cov-branch --cov-report=term-missing --cov-fail-under=85 -q
+      - name: CLI coverage (fail-under=60)
+        run: |
+          pytest tests/cli --cov=teds_core --cov=teds --cov-branch --cov-report=term-missing --cov-fail-under=60 -q
       - name: Verify demo/sample_tests.yaml (expect rc=1)
         run: |
           set +e
